requirements.txt
tqdm
torchvision
torch
numpy
pandas
pillow
pydensecrf
albumentations
opencv